Feature/auction (#13)
* fix: update http request and error response

* fix: add auto create auction if type = 1

* feat: add approve and reject auction

* feat: add get all bids from an auction

* feat: add image url to product list response

+++ API_v1/Controllers/ProductController.cs
@@ -3,10 +3,12 @@ using API.Request;
 using API.Response;
 using API.Response.ProductRes;
 using AutoMapper;
+using DataAccess;
 using DataAccess.Models;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Service;
+using System.Net;
 
 namespace API.Controllers
 {
@@ -14,21 +16,36 @@ namespace API.Controllers
     [ApiController]
     public class ProductController : ControllerBase
     {
+        private readonly IImageService _imageService;
         private readonly IProductService _productService;
+        private readonly IUserService _userService;
         private readonly IMapper _mapper;
 
-        public ProductController(IProductService productService, IMapper mapper)
+        public ProductController(IProductService productService, IUserService userService, IMapper mapper, IImageService imageService)
         {
             _productService = productService;
+            _userService = userService;
             _mapper = mapper;
+            _imageService = imageService;
+        }
+
+        private int GetUserIdFromToken()
+        {
+            var user = HttpContext.User;
+            return int.Parse(user.Claims.FirstOrDefault(p => p.Type == "UserId").Value);
         }
 
         [HttpGet]
-        public IActionResult GetProducts([FromQuery] string? nameSearch, [FromQuery] int materialId, [FromQuery] int categoryId, [FromQuery] decimal priceMin, [FromQuery] decimal priceMax, [FromQuery] int orderBy)
+        public IActionResult GetProducts([FromQuery] string? nameSearch, [FromQuery] int materialId, [FromQuery] int categoryId, [FromQuery] int type, [FromQuery] decimal priceMin, [FromQuery] decimal priceMax, [FromQuery] int orderBy)
         {
-            List<Product> productList = _productService.GetProducts(nameSearch, materialId, categoryId, priceMin, priceMax, orderBy);
-            List<ProductResponse> response = _mapper.Map<List<ProductResponse>>(productList);
-            return Ok(new BaseResponse { Code = 200, Message = "Get products successfully", Data = response });
+            List<Product> productList = _productService.GetProducts(nameSearch, materialId, categoryId, type, priceMin, priceMax, orderBy);
+            List<ProductListResponse> response = _mapper.Map<List<ProductListResponse>>(productList);
+            foreach (var item in response)
+            {
+                ProductImage image = _imageService.GetMainImageByProductId(item.Id);
+                item.ImageUrl = image.ImageUrl;
+            }
+            return Ok(new BaseResponse { Code = (int)HttpStatusCode.OK, Message = "Get products successfully", Data = response });
         }
 
         [HttpGet("seller/{id}")]
@@ -39,8 +56,13 @@ namespace API.Controllers
             {
                 return NotFound(new ErrorDetails { StatusCode = 400, Message = "This seller is not exist" });
             }
-            List<ProductResponse> response = _mapper.Map<List<ProductResponse>>(productList);
-            return Ok(new BaseResponse { Code = 200, Message = "Get products by seller successfully", Data = response });
+            List<ProductListResponse> response = _mapper.Map<List<ProductListResponse>>(productList);
+            foreach (var item in response)
+            {
+                ProductImage image = _imageService.GetMainImageByProductId(item.Id);
+                item.ImageUrl = image.ImageUrl;
+            }
+            return Ok(new BaseResponse { Code = (int)HttpStatusCode.OK, Message = "Get products by seller successfully", Data = response });
         }
 
         [HttpGet("{id}")]
@@ -52,36 +74,72 @@ namespace API.Controllers
                 return NotFound(new ErrorDetails { StatusCode = 400, Message = "This product is not exist" });
             }
             ProductResponse response = _mapper.Map<ProductResponse>(product);
-            return Ok(new BaseResponse { Code = 200, Message = "Get product successfully", Data = response });
+            return Ok(new BaseResponse { Code = (int)HttpStatusCode.OK, Message = "Get product successfully", Data = response });
         }
 
         [HttpPost]
         public IActionResult RegisterProduct([FromBody] ProductRequest request)
         {
-            _productService.CreateProduct(_mapper.Map<Product>(request));
-            return Ok(new BaseResponse { Code = 200, Message = "Register product successfully", Data = null });
+            //var userId = GetUserIdFromToken();
+            //var user = _userService.Get(userId);
+            //if (user == null || user.Role != (int) Role.Seller)
+            //{
+            //    return Unauthorized(new ErrorDetails { StatusCode = (int) HttpStatusCode.Unauthorized, Message = "You are not allowed to access this" });
+            //}
+
+            var auctionRequest = new AuctionRequest
+            {
+                Title = request.Title,
+                Step = request.Step,
+                StartedAt = request.StartedAt,
+                EndedAt = request.EndedAt,
+            };
+
+            _productService.CreateProduct(_mapper.Map<Product>(request), _mapper.Map<Auction>(auctionRequest));
+            return Ok(new BaseResponse { Code = (int)HttpStatusCode.OK, Message = "Register product successfully", Data = null });
         }
 
         [HttpPatch("{id}")]
         public IActionResult EditProduct(int id, [FromBody] ProductRequest request)
         {
-            Product product = _productService.UpdateProduct(id, _mapper.Map<Product>(request));
+            //var userId = GetUserIdFromToken();
+            //var user = _userService.Get(userId);
+            //if (user == null || user.Role != (int)Role.Seller)
+            //{
+            //    return Unauthorized(new ErrorDetails { StatusCode = (int)HttpStatusCode.Unauthorized, Message = "You are not allowed to access this" });
+            //}
+
+            var auctionRequest = new AuctionRequest
+            {
+                Title = request.Title,
+                Step = request.Step,
+                StartedAt = request.StartedAt,
+                EndedAt = request.EndedAt,
+            };
+
+            Product product = _productService.UpdateProduct(id, _mapper.Map<Product>(request), _mapper.Map<Auction>(auctionRequest));
             if (product == null)
             {
                 return NotFound(new ErrorDetails { StatusCode = 400, Message = "This product is not exist" });
             }
-            return Ok(new BaseResponse { Code = 200, Message = "Edit product successfully", Data = null });
+            return Ok(new BaseResponse { Code = (int)HttpStatusCode.OK, Message = "Edit product successfully", Data = null });
         }
 
         [HttpDelete("{id}")]
         public IActionResult DeleteProduct(int id)
         {
+            var userId = GetUserIdFromToken();
+            var user = _userService.Get(userId);
+            if (user == null || user.Role != (int)Role.Seller)
+            {
+                return Unauthorized(new ErrorDetails { StatusCode = (int)HttpStatusCode.Unauthorized, Message = "You are not allowed to access this" });
+            }
             Product product = _productService.DeleteProduct(id);
             if (product == null)
             {
                 return NotFound(new ErrorDetails { StatusCode = 400, Message = "This product is not exist" });
             }
-            return Ok(new BaseResponse { Code = 200, Message = "Edit product successfully", Data = null });
+            return Ok(new BaseResponse { Code = (int)HttpStatusCode.OK, Message = "Edit product successfully", Data = null });
         }
     }
 }
